+++ src/pages/signup.tsx
@@ -1,0 +1,5 @@
+import SignUp from "@/components/signup";
+
+export default function LoginPage() {
+  return <SignUp />;
+}
